requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pymongo==4.2.0
pyparsing==3.0.9
pyTelegramBotAPI==4.7.0
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
soupsieve==2.3.2.post1
tabulate==0.9.0
Telethon==1.25.1
tomli==2.0.1
urllib3==1.26.12
python-telegram-bot==13.14
